split cspace to a crate
