+++ .gitignore
@@ -12,9 +12,11 @@ Cargo.lock
 # MSVC Windows builds of rustc generate these, which store debugging information
 *.pdb
 .idea/
+.vscode/
 
 
 *build/
+*target/
 
 # Added by cargo
 #
