+++ common/src/lib.rs
@@ -1,0 +1,9 @@
+#![no_std]
+#![allow(dead_code)]
+#![allow(non_snake_case)]
+#![allow(non_camel_case_types)]
+#![allow(non_upper_case_globals)]
+
+pub mod sel4_config;
+pub mod structures;
+mod utils;
