+++ common/src/utils.rs
@@ -1,0 +1,17 @@
+#[macro_export]
+macro_rules! MASK {
+    ($e:expr) => {
+        {
+             (1usize << $e) - 1usize
+        }
+    }
+}
+
+#[macro_export]
+macro_rules! BIT {
+    ($e:expr) => {
+        {
+            1usize<<$e
+        }
+    }
+}
