+++ cspace/src/cap/domain.rs
