+++ cspace/src/lang_items.rs
@@ -1,0 +1,6 @@
+use core::panic::PanicInfo;
+
+#[panic_handler]
+fn panic(info: &PanicInfo) -> ! {
+    panic!("cspace error!")
+}
