+++ cspace/src/utils.rs
@@ -1,0 +1,51 @@
+use common::sel4_config::{RISCVPageBits, RISCVGigaPageBits, RISCVMegaPageBits, seL4_MinUntypedBits,
+    RISCV_4K_Page, RISCV_Mega_Page, RISCV_Giga_Page};
+
+#[macro_export]
+macro_rules! MASK {
+    ($e:expr) => {
+        {
+             (1usize << $e) - 1usize
+        }
+    }
+}
+
+#[macro_export]
+macro_rules! BIT {
+    ($e:expr) => {
+        {
+            1usize<<$e
+        }
+    }
+}
+
+#[inline]
+pub fn pageBitsForSize(page_size: usize) -> usize {
+    match page_size {
+        RISCV_4K_Page => RISCVPageBits,
+        RISCV_Mega_Page => RISCVMegaPageBits,
+        RISCV_Giga_Page => RISCVGigaPageBits,
+        _ => panic!("Invalid page size!"),
+    }
+}
+
+#[inline]
+pub fn convert_to_type_ref<T>(addr: usize) -> &'static T {
+    assert_ne!(addr, 0);
+    unsafe {
+        & *(addr as *mut T)
+    }
+}
+
+#[inline]
+pub fn convert_to_mut_type_ref<T>(addr: usize) -> &'static mut T {
+    assert_ne!(addr, 0);
+    unsafe {
+        &mut *(addr as *mut T)
+    }
+}
+
+
+pub fn MAX_FREE_INDEX(bits: usize) -> usize {
+    BIT!(bits - seL4_MinUntypedBits)
+}
